+++ float64validator/between_test.go
@@ -5,13 +5,14 @@ package float64validator_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-framework-validators/float64validator"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-
-	"github.com/hashicorp/terraform-plugin-framework-validators/float64validator"
 )
 
 func TestBetweenValidator(t *testing.T) {
@@ -66,11 +67,18 @@ func TestBetweenValidator(t *testing.T) {
 			max:         3.10,
 			expectError: true,
 		},
+		"invalid validator usage - minVal > maxVal": {
+			val:         types.Float64Value(2),
+			min:         3.20,
+			max:         3.10,
+			expectError: true,
+		},
 	}
 
 	for name, test := range tests {
 		name, test := name, test
-		t.Run(name, func(t *testing.T) {
+
+		t.Run(fmt.Sprintf("ValidateFloat64 - %s", name), func(t *testing.T) {
 			t.Parallel()
 			request := validator.Float64Request{
 				Path:           path.Root("test"),
@@ -86,6 +94,24 @@ func TestBetweenValidator(t *testing.T) {
 
 			if response.Diagnostics.HasError() && !test.expectError {
 				t.Fatalf("got unexpected error: %s", response.Diagnostics)
+			}
+		})
+
+		t.Run(fmt.Sprintf("ValidateParameterFloat64 - %s", name), func(t *testing.T) {
+			t.Parallel()
+			request := function.Float64ParameterValidatorRequest{
+				ArgumentPosition: 0,
+				Value:            test.val,
+			}
+			response := function.Float64ParameterValidatorResponse{}
+			float64validator.Between(test.min, test.max).ValidateParameterFloat64(context.TODO(), request, &response)
+
+			if response.Error == nil && test.expectError {
+				t.Fatal("expected error, got no error")
+			}
+
+			if response.Error != nil && !test.expectError {
+				t.Fatalf("got unexpected error: %s", response.Error)
 			}
 		})
 	}