+++ int32validator/one_of_test.go
@@ -5,83 +5,95 @@ package int32validator_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-framework-validators/int32validator"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-
-	"github.com/hashicorp/terraform-plugin-framework-validators/int32validator"
 )
 
 func TestOneOfValidator(t *testing.T) {
 	t.Parallel()
 
 	type testCase struct {
-		in        types.Int32
-		validator validator.Int32
-		expErrors int
+		in          types.Int32
+		oneOfValues []int32
+		expectError bool
 	}
 
 	testCases := map[string]testCase{
 		"simple-match": {
 			in: types.Int32Value(123),
-			validator: int32validator.OneOf(
+			oneOfValues: []int32{
 				123,
 				234,
 				8910,
 				1213,
-			),
-			expErrors: 0,
+			},
 		},
 		"simple-mismatch": {
 			in: types.Int32Value(123),
-			validator: int32validator.OneOf(
+			oneOfValues: []int32{
 				234,
 				8910,
 				1213,
-			),
-			expErrors: 1,
+			},
+			expectError: true,
 		},
 		"skip-validation-on-null": {
 			in: types.Int32Null(),
-			validator: int32validator.OneOf(
+			oneOfValues: []int32{
 				234,
 				8910,
 				1213,
-			),
-			expErrors: 0,
+			},
 		},
 		"skip-validation-on-unknown": {
 			in: types.Int32Unknown(),
-			validator: int32validator.OneOf(
+			oneOfValues: []int32{
 				234,
 				8910,
 				1213,
-			),
-			expErrors: 0,
+			},
 		},
 	}
 
 	for name, test := range testCases {
 		name, test := name, test
-		t.Run(name, func(t *testing.T) {
+
+		t.Run(fmt.Sprintf("ValidateInt32 - %s", name), func(t *testing.T) {
 			t.Parallel()
 			req := validator.Int32Request{
 				ConfigValue: test.in,
 			}
 			res := validator.Int32Response{}
-			test.validator.ValidateInt32(context.TODO(), req, &res)
+			int32validator.OneOf(test.oneOfValues...).ValidateInt32(context.TODO(), req, &res)
 
-			if test.expErrors > 0 && !res.Diagnostics.HasError() {
-				t.Fatalf("expected %d error(s), got none", test.expErrors)
+			if !res.Diagnostics.HasError() && test.expectError {
+				t.Fatal("expected error, got no error")
 			}
 
-			if test.expErrors > 0 && test.expErrors != res.Diagnostics.ErrorsCount() {
-				t.Fatalf("expected %d error(s), got %d: %v", test.expErrors, res.Diagnostics.ErrorsCount(), res.Diagnostics)
+			if res.Diagnostics.HasError() && !test.expectError {
+				t.Fatalf("got unexpected error: %s", res.Diagnostics)
+			}
+		})
+
+		t.Run(fmt.Sprintf("ValidateParameterInt32 - %s", name), func(t *testing.T) {
+			t.Parallel()
+			req := function.Int32ParameterValidatorRequest{
+				Value: test.in,
+			}
+			res := function.Int32ParameterValidatorResponse{}
+			int32validator.OneOf(test.oneOfValues...).ValidateParameterInt32(context.TODO(), req, &res)
+
+			if res.Error == nil && test.expectError {
+				t.Fatal("expected error, got no error")
 			}
 
-			if test.expErrors == 0 && res.Diagnostics.HasError() {
-				t.Fatalf("expected no error(s), got %d: %v", res.Diagnostics.ErrorsCount(), res.Diagnostics)
+			if res.Error != nil && !test.expectError {
+				t.Fatalf("got unexpected error: %s", res.Error)
 			}
 		})
 	}