+++ int64validator/at_least_example_test.go
@@ -6,6 +6,7 @@ package int64validator_test
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
 
@@ -16,6 +17,20 @@ func ExampleAtLeast() {
 			"example_attr": schema.Int64Attribute{
 				Required: true,
 				Validators: []validator.Int64{
+					// Validate integer value must be at least 42
+					int64validator.AtLeast(42),
+				},
+			},
+		},
+	}
+}
+
+func ExampleAtLeast_function() {
+	_ = function.Definition{
+		Parameters: []function.Parameter{
+			function.Int64Parameter{
+				Name: "example_param",
+				Validators: []function.Int64ParameterValidator{
 					// Validate integer value must be at least 42
 					int64validator.AtLeast(42),
 				},