+++ internal/configvalidator/conflicting_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
+	"github.com/hashicorp/terraform-plugin-framework/ephemeral"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -934,6 +935,115 @@ func TestConflictingValidatorValidateResource(t *testing.T) {
 			got := &resource.ValidateConfigResponse{}
 
 			testCase.validator.ValidateResource(context.Background(), testCase.req, got)
+
+			if diff := cmp.Diff(got, testCase.expected); diff != "" {
+				t.Errorf("unexpected difference: %s", diff)
+			}
+		})
+	}
+}
+
+func TestConflictingValidatorValidateEphemeralResource(t *testing.T) {
+	t.Parallel()
+
+	testCases := map[string]struct {
+		validator configvalidator.ConflictingValidator
+		req       ephemeral.ValidateConfigRequest
+		expected  *ephemeral.ValidateConfigResponse
+	}{
+		"no-diagnostics": {
+			validator: configvalidator.ConflictingValidator{
+				PathExpressions: path.Expressions{
+					path.MatchRoot("test"),
+				},
+			},
+			req: ephemeral.ValidateConfigRequest{
+				Config: tfsdk.Config{
+					Schema: schema.Schema{
+						Attributes: map[string]schema.Attribute{
+							"test": schema.StringAttribute{
+								Optional: true,
+							},
+							"other": schema.StringAttribute{
+								Optional: true,
+							},
+						},
+					},
+					Raw: tftypes.NewValue(
+						tftypes.Object{
+							AttributeTypes: map[string]tftypes.Type{
+								"test":  tftypes.String,
+								"other": tftypes.String,
+							},
+						},
+						map[string]tftypes.Value{
+							"test":  tftypes.NewValue(tftypes.String, "test-value"),
+							"other": tftypes.NewValue(tftypes.String, "test-value"),
+						},
+					),
+				},
+			},
+			expected: &ephemeral.ValidateConfigResponse{},
+		},
+		"diagnostics": {
+			validator: configvalidator.ConflictingValidator{
+				PathExpressions: path.Expressions{
+					path.MatchRoot("test1"),
+					path.MatchRoot("test2"),
+				},
+			},
+			req: ephemeral.ValidateConfigRequest{
+				Config: tfsdk.Config{
+					Schema: schema.Schema{
+						Attributes: map[string]schema.Attribute{
+							"test1": schema.StringAttribute{
+								Optional: true,
+							},
+							"test2": schema.StringAttribute{
+								Optional: true,
+							},
+							"other": schema.StringAttribute{
+								Optional: true,
+							},
+						},
+					},
+					Raw: tftypes.NewValue(
+						tftypes.Object{
+							AttributeTypes: map[string]tftypes.Type{
+								"test1": tftypes.String,
+								"test2": tftypes.String,
+								"other": tftypes.String,
+							},
+						},
+						map[string]tftypes.Value{
+							"test1": tftypes.NewValue(tftypes.String, "test-value"),
+							"test2": tftypes.NewValue(tftypes.String, "test-value"),
+							"other": tftypes.NewValue(tftypes.String, "test-value"),
+						},
+					),
+				},
+			},
+			expected: &ephemeral.ValidateConfigResponse{
+				Diagnostics: diag.Diagnostics{
+					diag.NewAttributeErrorDiagnostic(
+						path.Root("test1"),
+						"Invalid Attribute Combination",
+						"These attributes cannot be configured together: [test1,test2]",
+					),
+				},
+			},
+		},
+	}
+
+	for name, testCase := range testCases {
+		name, testCase := name, testCase
+
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+
+			got := &ephemeral.ValidateConfigResponse{}
+
+			testCase.validator.ValidateEphemeralResource(context.Background(), testCase.req, got)
 
 			if diff := cmp.Diff(got, testCase.expected); diff != "" {
 				t.Errorf("unexpected difference: %s", diff)