+++ internal/schemavalidator/at_least_one_of_test.go
@@ -84,6 +84,34 @@ func TestAtLeastOneOfValidatorValidate(t *testing.T) {
 			},
 			expected: &schemavalidator.AtLeastOneOfValidatorResponse{},
 		},
+		"self-is-unknown": {
+			req: schemavalidator.AtLeastOneOfValidatorRequest{
+				ConfigValue:    types.StringUnknown(),
+				Path:           path.Root("bar"),
+				PathExpression: path.MatchRoot("bar"),
+				Config: tfsdk.Config{
+					Schema: schema.Schema{
+						Attributes: map[string]schema.Attribute{
+							"foo": schema.Int64Attribute{},
+							"bar": schema.StringAttribute{},
+						},
+					},
+					Raw: tftypes.NewValue(tftypes.Object{
+						AttributeTypes: map[string]tftypes.Type{
+							"foo": tftypes.Number,
+							"bar": tftypes.String,
+						},
+					}, map[string]tftypes.Value{
+						"foo": tftypes.NewValue(tftypes.Number, 42),
+						"bar": tftypes.NewValue(tftypes.String, tftypes.UnknownValue),
+					}),
+				},
+			},
+			in: path.Expressions{
+				path.MatchRoot("foo"),
+			},
+			expected: &schemavalidator.AtLeastOneOfValidatorResponse{},
+		},
 		"error_none-set": {
 			req: schemavalidator.AtLeastOneOfValidatorRequest{
 				ConfigValue:    types.StringNull(),