+++ mapvalidator/size_at_least_example_test.go
@@ -6,6 +6,7 @@ package mapvalidator_test
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/mapvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
@@ -18,6 +19,20 @@ func ExampleSizeAtLeast() {
 				ElementType: types.StringType,
 				Required:    true,
 				Validators: []validator.Map{
+					// Validate this map must contain at least 2 elements.
+					mapvalidator.SizeAtLeast(2),
+				},
+			},
+		},
+	}
+}
+
+func ExampleSizeAtLeast_function() {
+	_ = function.Definition{
+		Parameters: []function.Parameter{
+			function.MapParameter{
+				Name: "example_param",
+				Validators: []function.MapParameterValidator{
 					// Validate this map must contain at least 2 elements.
 					mapvalidator.SizeAtLeast(2),
 				},