+++ setvalidator/size_between_example_test.go
@@ -6,6 +6,7 @@ package setvalidator_test
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/setvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
@@ -18,6 +19,20 @@ func ExampleSizeBetween() {
 				ElementType: types.StringType,
 				Required:    true,
 				Validators: []validator.Set{
+					// Validate this set must contain at least 2 and at most 4 elements.
+					setvalidator.SizeBetween(2, 4),
+				},
+			},
+		},
+	}
+}
+
+func ExampleSizeBetween_function() {
+	_ = function.Definition{
+		Parameters: []function.Parameter{
+			function.SetParameter{
+				Name: "example_param",
+				Validators: []function.SetParameterValidator{
 					// Validate this set must contain at least 2 and at most 4 elements.
 					setvalidator.SizeBetween(2, 4),
 				},