+++ stringvalidator/utf8_length_at_most_example_test.go
@@ -6,6 +6,7 @@ package stringvalidator_test
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
 
@@ -16,6 +17,20 @@ func ExampleUTF8LengthAtMost() {
 			"example_attr": schema.StringAttribute{
 				Required: true,
 				Validators: []validator.String{
+					// Validate UTF-8 character count must be at most 255 characters.
+					stringvalidator.UTF8LengthAtMost(255),
+				},
+			},
+		},
+	}
+}
+
+func ExampleUTF8LengthAtMost_function() {
+	_ = function.Definition{
+		Parameters: []function.Parameter{
+			function.StringParameter{
+				Name: "example_param",
+				Validators: []function.StringParameterValidator{
 					// Validate UTF-8 character count must be at most 255 characters.
 					stringvalidator.UTF8LengthAtMost(255),
 				},